[Anmol Singh]Extension: getFormattedDate() extended in Date

+++ CO777245_W2020_MID_TERM_MAD3004/CO777245_W2020_MID_TERM_MAD3004/Double+extension.swift
@@ -11,7 +11,7 @@ import Foundation
    extension Double{
        func getFormattedAmount() -> String
     {
-           return String(format: "$%0.2f", Float(self))
+           return String(format: "$%0.2f", Double(self))
     }
